requirements.txt
Django==5.0.3
environs==11.0.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.31
pillow==10.2.0